Prepared response for GetMessage method
