+++ src/main/java/ru/scarletredman/gd2spring/controller/response/GetMessagesResponse.java
@@ -1,13 +1,11 @@
 package ru.scarletredman.gd2spring.controller.response;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import java.util.EnumMap;
-import java.util.Map;
+import java.util.LinkedList;
+import java.util.List;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import ru.scarletredman.gd2spring.controller.response.json.ResponseSerializer;
-import ru.scarletredman.gd2spring.util.JoinResponseUtil;
 
 @Getter
 @Setter
@@ -16,7 +14,7 @@ public class GetMessagesResponse implements ResponseSerializer.Response {
 
     private long messagesCount;
     private long offset;
-    private final Map<Key, Object> elements = new EnumMap<>(Key.class);
+    private final List<MessageResponse> messages = new LinkedList<>();
 
     public GetMessagesResponse(long messagesCount, long offset) {
         this.messagesCount = messagesCount;
@@ -30,22 +28,8 @@ public class GetMessagesResponse implements ResponseSerializer.Response {
 
     @Override
     public String getResponse() {
-        return JoinResponseUtil.join(elements, "|") + "#" + messagesCount + ":" + offset + ":10";
-    }
-
-    @Getter
-    @RequiredArgsConstructor
-    public enum Key implements JoinResponseUtil.Key {
-        MESSAGE_ID("1"),
-        USER_ID1("2"),
-        USER_ID2("3"),
-        SUBJECT("4"),
-        USERNAME("6"),
-        UPLOAD_TIME("7"),
-        IS_NEW("8"),
-        GET_SENT("9"),
-        ;
-
-        private final String code;
+        var msgList = String.join(
+                "|", messages.stream().map(MessageResponse::getResponse).toList());
+        return msgList + "#" + messagesCount + ":" + offset + ":10";
     }
 }
